Updated text field to use a controller
